+++ repository/appStyleRepository.go
@@ -7,7 +7,7 @@ import (
 )
 
 // UpdateProfileBanner 更新用户页banner
-func UpdateProfileBanner(uid uint, bannerName string) error {
+func UpdateProfileBanner(uid int64, bannerName string) error {
 	style := &styleModel.AppStyle{Uid: uid, ProfileBanner: bannerName}
 	result := global.Db.Model(&styleModel.AppStyle{}).Where("uid = ?", uid).Updates(map[string]interface{}{"uid": style.Uid, "profile_banner": style.ProfileBanner})
 	if result.RowsAffected == 0 {
@@ -20,7 +20,7 @@ func UpdateProfileBanner(uid uint, bannerName string) error {
 }
 
 // GetLastBanner 获取用户原始banner图片地址
-func GetLastBanner(uid uint) (string, error) {
+func GetLastBanner(uid int64) (string, error) {
 	var style *styleModel.AppStyle
 	if err := global.Db.Where("uid = ?", uid).Select("profile_banner").First(&style).Error; err != nil {
 		return "", err
@@ -29,7 +29,7 @@ func GetLastBanner(uid uint) (string, error) {
 }
 
 // GetStyle 获取app风格数据
-func GetStyle(uid uint) (*styleModel.AppStyle, error) {
+func GetStyle(uid int64) (*styleModel.AppStyle, error) {
 	var style *styleModel.AppStyle
 	if err := global.Db.Where("uid = ?", uid).First(&style).Error; err != nil {
 		return nil, err
